+++ recommends/tests/__init__.py
@@ -1,0 +1,2 @@
+#from recommends.tests.tests import *
+from recommends.tests.providers import *